[DO NOT MERGE] Fixed merge conflicts on my end. Fixed edge constructor. getDistance should return a double unless there's a specific reason it needs to return an int. Made a small change to the tests for AbsNode. Do not merge from my branch unless you know all of the files committed are up to date with yours.

+++ src/node/Edge.java
@@ -8,7 +8,7 @@ public class Edge {
 	private AbsNode to;
 	private double distance;
 	
-	public Edge(AbsNode fromNode, AbsNode toNode, int dist){
+	public Edge(AbsNode fromNode, AbsNode toNode, double dist){
 		from = fromNode;
 		to = toNode;
 		distance = dist;
